requirements.txt
alembic==1.4.3
boto3==1.17.3
botocore==1.20.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==8.0.1
python-dateutil==2.8.1
python-editor==1.0.4
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.3
Werkzeug==1.0.1

# Requirements for Heroku
psycopg2==2.7.3.1
gunicorn==19.7.1